% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{import_libre_franklin}
\alias{import_libre_franklin}
\title{Import Libre Franklin font for use in charts}
\usage{
import_libre_franklin()
}
\description{
Libre Franklin Libre Franklin is an interpretation and expansion of the 1912 Morris Fuller Benton classic.
The Libre Franklin project is led by Impallari Type, a type design foundry based in Rosario, Argentina.
To contribute, see github.com/impallari/
}
\details{
Libre-Franklin is Copyright 2015 Impallari Type and are licensed under the Open Font License.
}
\note{
This will take care of ensuring PDF/PostScript usage. The location of the
font directory is displayed after the base import is complete. It is highly
recommended that you install them on your system the same way you would any
other font you wish to use in other programs.
}
