% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\docType{data}
\name{font_fl}
\alias{font_fl}
\alias{font_fl_black}
\alias{font_fl_bold}
\alias{font_fl_light}
\alias{font_fl_thin}
\title{Libre Franklin font name R variable aliases}
\format{
length 1 character vector

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
font_fl

font_fl_black

font_fl_bold

font_fl_light

font_fl_thin
}
\description{
\code{font_tw} == "\verb{Libre Franklin}"

\code{font_fl_black} == "\code{LibreFranklin-Black}"

\code{font_fl_bold} == "\code{LibreFranklin-Bold}"

\code{font_fl_light} == "\code{LibreFranklin-Light}"

\code{font_fl_thin} == "\code{LibreFranklin-Thin}"
}
\note{
\code{font_fl_black} (a.k.a. "\code{LibreFranklin-Black}") is not available on
Windows and will throw a warning if used in plots.

\code{font_fl_bold} (a.k.a. "\verb{Libre Franklin Bold}") is not available on
Windows and will throw a warning if used in plots.
}
\keyword{datasets}
